requirements.txt
psycopg2-binary==2.9.9
pymongo==4.7.2
faker==25.2.0
matplotlib==3.9.0
seaborn==0.13.2
numpy==1.26.4
